recaftor wrgdv3: rename ColumnDefinition to makeColumnDefinition
