+++ src/scripts/modules/gooddata-writer-v3/helpers/makeColumnDefinition.js
